+++ src/BookPetVersion.h
@@ -1,0 +1,7 @@
+#pragma once
+
+inline constexpr char kBookPetVersion[] = "0.5.0";
+
+#ifndef BOOKPET_VERSION
+#define BOOKPET_VERSION kBookPetVersion
+#endif